@@ -9,8 +9,8 @@
     "@types/react-dom": "^16.0.7",
     "babel-preset-react": "^6.24.1",
     "prop-types": "^15.6.2",
-    "react": "^16.4.2",
-    "react-dom": "^16.4.2",
+    "react": "^16.5.2",
+    "react-dom": "^16.5.2",
     "react-player": "^1.6.4",
     "react_ujs": "^2.4.4",
     "semantic-ui-react": "^0.82.2",
